+++ docs/generate_precomputed_list.py
@@ -13,21 +13,21 @@ def generate_documentation_table():
 
     data_ = []
     for info in lib.get_models_info():
-        kind = info.model['kind']
-        features = ' '.join(info.model['feature_names'])
-        extc = info.extinction['source']
-        atm = info.atmosphere['source']
-        teffmin = info.atmosphere['teff'][0]
-        teffmax = info.atmosphere['teff'][1]
-        loggmin = info.atmosphere['logg'][0]
-        loggmax = info.atmosphere['logg'][1]
-        fehmin = info.atmosphere['feh'][0]
-        fehmax = info.atmosphere['feh'][1]
-        a0min = info.extinction['A0'][0]
-        a0max = info.extinction['A0'][1]
-        if 'R0' in info.extinction:
-            r0min = info.extinction['R0'][0]
-            r0max = info.extinction['R0'][1]
+        kind = info.model["kind"]
+        features = " ".join(info.model["feature_names"])
+        extc = info.extinction["source"]
+        atm = info.atmosphere["source"]
+        teffmin = info.atmosphere["teff"][0]
+        teffmax = info.atmosphere["teff"][1]
+        loggmin = info.atmosphere["logg"][0]
+        loggmax = info.atmosphere["logg"][1]
+        fehmin = info.atmosphere["feh"][0]
+        fehmax = info.atmosphere["feh"][1]
+        a0min = info.extinction["A0"][0]
+        a0max = info.extinction["A0"][1]
+        if "R0" in info.extinction:
+            r0min = info.extinction["R0"][0]
+            r0max = info.extinction["R0"][1]
         else:
             r0min, r0max = 3.1, 3.1