+++ src/dustapprox/tools/parallel.py
@@ -15,8 +15,11 @@ def tqdm_joblib(tqdm_object: tqdm):
 
         import time
         from joblib import Parallel, delayed
+
+
         def some_method(wait_time):
             time.sleep(wait_time)
+
 
         with tqdm_joblib(tqdm(desc="My method", total=10)) as progress_bar:
             Parallel(n_jobs=2)(delayed(some_method)(0.2) for i in range(10))